#24 #33 rework of favorite trips and error messages

+++ app/controllers/api/v1/trips_controller.rb
@@ -33,42 +33,34 @@ class Api::V1::TripsController < ApplicationController
              include: { trip_place_infos: { include: { place: { include: [:address, :category_dictionaries] } } } },
              methods: [:favorite]
     else
-      render json: { error: @trip.errors.full_messages }, status: :unprocessable_entity
+      render json: { error_message: @trip.errors.full_messages, error_code: 422 }, status: :unprocessable_entity
     end
   end
 
   # PATCH/PUT /api/v1/trips/1
   def update
     Trip.current_user = User.find(doorkeeper_token.resource_owner_id)
-    if @trip.update(trip_params)
+    favorite = trip_params[:favorite]
+    uft_options = { trip_id: @trip.id, user_id: doorkeeper_token.resource_owner_id }
+    uft = UserFavoriteTrip.find_by(uft_options)
+    if favorite == 'true' || favorite == true
+      UserFavoriteTrip.where(uft_options).first_or_create unless uft.present?
+    elsif favorite == 'false' || favorite == false
+      uft.destroy if uft.present?
+    end
+
+    if @trip.update(trip_params.except(:favorite))
       render json: @trip,
              include: { trip_place_infos: { include: { place: { include: [:address, :category_dictionaries] } } } },
              methods: [:favorite]
     else
-      render json: { error: @trip.errors.full_messages }, status: :unprocessable_entity
+      render json: { error_message: @trip.errors.full_messages, error_code: 422 }, status: :unprocessable_entity
     end
   end
 
   # DELETE /api/v1/trips/1
   def destroy
     @trip.destroy
-  end
-
-  def add_to_favorite
-    UserFavoriteTrip.where(trip_id: @trip.id, user_id: doorkeeper_token.resource_owner_id).first_or_create
-
-    render json: { success: true }
-  end
-
-  def remove_from_favorite
-    uft = UserFavoriteTrip.where(trip_id: @trip.id, user_id: doorkeeper_token.resource_owner_id).first
-
-    if uft
-      uft.destroy
-      render json: { success: true }
-    else
-      render json: { error: "Trip #{@trip.id} was not added to favorite by user #{doorkeeper_token.resource_owner_id}" }
-    end
   end
 
   private
@@ -80,6 +72,6 @@ class Api::V1::TripsController < ApplicationController
 
   # Only allow a list of trusted parameters through.
   def trip_params
-    params.require(:trip).permit(:name, :description, :distance, :duration, :user_id, :image_url)
+    params.require(:trip).permit(:name, :description, :distance, :duration, :user_id, :image_url, :favorite)
   end
 end
